pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "nes"
version = "0.1.0"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

[project.optional-dependencies]
rlcade = [
    "gymnasium>=1.0",
    "numpy>=1.26,<2",
    "torch>=2.10",
    "tqdm>=4.60",
    "tensorboard>=2.14",
    "viztracer>=0.16",
]
notebook = [
    "ipython>=8.0",
    "ipywidgets>=8.0",
    "Pillow>=9.0",
]
ray = [
    "ray[default]>=2.9",
]
cuda = [
    "nvidia-ml-py>=12.0",
]
s3 = [
    "boto3>=1.28",
    "s3torchconnector>=1.0",
]
all = ["nes[rlcade,notebook,ray,cuda,s3]"]

[tool.maturin]
manifest-path = "crates/rlcade/Cargo.toml"
module-name = "nes"
python-packages = ["rlcade"]

[tool.black]
line-length = 120
target-version = ["py312"]
